apiextensions: move ObjectMeta coercion into own package
